salary is now BigDecimal

+++ src/main/java/ru/tsconsulting/controllers/EmployeesController.java
@@ -22,6 +22,7 @@ import ru.tsconsulting.repositories.GradeRepository;
 import ru.tsconsulting.repositories.PositionRepository;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.*;
@@ -126,7 +127,7 @@ public class EmployeesController {
                                  @ApiParam(value = "Id of new grade")
                                  @RequestParam(value = "newGrade", required = false) Long newGrade,
                                  @ApiParam(value = "New salary")
-                                 @RequestParam(value = "newSalary", required = false) Long newSalary,
+                                 @RequestParam(value = "newSalary", required = false) BigDecimal newSalary,
                                  HttpServletRequest request) {
         Employee employee = employeeRepository.findById(employeeId);
         if (employee == null) {
